fix: refactor to make linter happy
